requirements.txt
pyspark==3.5.1
psycopg2-binary
python-bcb 
pytest
pandas
pytest-mock
